Update README.md
Added a Q&A for partners, as follows: 
What action is required for IBM Business Partners and Intellectual Property Partners (IPPs)?
IBM Business Partners and Intellectual Property Partners (IPPs) are not contractually required to remove biased terminology. However, IBM encourages our partners to review their terminology on a voluntary basis to ensure that the language used in products and content supports racial and cultural inclusivity. Many organizations across the IT industry are engaging in this important movement to replace biased terminology.